Refactor grid_janitor to use new mig/lib/ for all helper functions and make the daemon itself very shallow.
